feat(reviews): permitir avaliacao em passos de meia estrela
Fecha #301: nota de review passava de SMALLINT (1 a 5 inteiro) pra
NUMERIC(2,1), com CHECK garantindo passos de 0.5 (rejeita algo como
3.3). Frontend: StarRating agora tem duas zonas de clique por estrela
(metade esquerda = X.5, direita = X.0) e o preenchimento visual usa
clipPath SVG em vez de liga/desliga por estrela inteira. Estrelas
tambem aumentadas de tamanho (22px -> 36px no modo interativo), que
estavam pequenas demais pra acertar a metade certa.

+++ backend/src/main/java/gerenciador_musica_backend/dto/ReviewAtualizacaoRequestDTO.java
@@ -1,7 +1,9 @@
 package gerenciador_musica_backend.dto;
 
+import java.math.BigDecimal;
+
 public record ReviewAtualizacaoRequestDTO(
-        Short nota,
+        BigDecimal nota,
         String texto
 ) {
 }
