+++ backend/src/main/java/gerenciador_musica_backend/dto/ReviewRequestDTO.java
@@ -1,9 +1,11 @@
 package gerenciador_musica_backend.dto;
 
+import java.math.BigDecimal;
+
 public record ReviewRequestDTO(
         Long idMusica,
         Long idAlbum,
-        Short nota,
+        BigDecimal nota,
         String texto
 ) {
 }
